Move Client models
