+++ src/app/clients/clients.repository.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { plainToInstance } from 'class-transformer';
 import { DatabaseService } from '../database';
-import { Client, CreateClientData } from './models';
+import { Client, CreateClientData } from '../shared';
 
 const selectedClientAttributes = {
     select: {
