+++ src/app/shared/index.ts
@@ -1,3 +1,4 @@
 export * from './authentication.models';
+export * from './client.models';
 export * from './token.models';
 export * from './user.models';
